+++ src/handlers/favorites_answer.py
@@ -33,9 +33,9 @@ from app_types.supports_bool import SupportsBool
 from app_types.update import Update
 from integrations.tg.tg_answers import TgAnswer, TgAnswerToSender, TgMessageAnswer, TgTextAnswer
 from services.answers.change_state_answer import ChangeStateAnswer
-from services.user_state import UserStep
 from srv.ayats.favorite_ayat_answer import FavoriteAyatAnswer
 from srv.ayats.favorite_ayat_empty_safe import FavoriteAyatEmptySafeAnswer
+from srv.users.user_step import UserStep
 
 
 @final
