+++ src/handlers/search_city_answer.py
@@ -40,12 +40,13 @@ from integrations.tg.tg_answers.interface import TgAnswer
 from integrations.tg.tg_answers.location_answer import TgLocationAnswer
 from integrations.tg.tg_answers.skip_not_processable import TgSkipNotProcessable
 from services.reset_state_answer import ResetStateAnswer
-from services.user_state import CachedUserState, RedisUserState
 from srv.prayers.change_city_answer import ChangeCityAnswer, CityNotSupportedAnswer
 from srv.prayers.city import PgCity
 from srv.prayers.update_user_city import PgUpdatedUserCity
 from srv.prayers.user_not_registered_safe_answer import UserNotRegisteredSafeAnswer
+from srv.users.cached_user_state import CachedUserState
 from srv.users.new_user import PgNewUser
+from srv.users.redis_user_state import RedisUserState
 
 
 @final
