+++ src/services/answers/change_state_answer.py
@@ -31,7 +31,8 @@ from app_types.logger import LogSink
 from app_types.update import Update
 from integrations.tg.chat_id import TgChatId
 from integrations.tg.tg_answers import TgAnswer
-from services.user_state import RedisUserState, UserStep
+from srv.users.redis_user_state import RedisUserState
+from srv.users.user_step import UserStep
 
 
 @final
