+++ src/srv/users/cached_user_state.py
@@ -22,37 +22,42 @@
 
 from typing import final, override
 
-import attrs
-import httpx
 from pyeo import elegant
-from redis.asyncio import Redis
 
-from app_types.logger import LogSink
-from app_types.update import Update
-from integrations.tg.chat_id import TgChatId
-from integrations.tg.tg_answers import TgAnswer
-from srv.users.redis_user_state import RedisUserState
+from srv.users.user_state import UserState
+from srv.users.user_step import UserStep
 
 
 @final
-@attrs.define(frozen=True)
 @elegant
-class StepAnswer(TgAnswer):
-    """Роутинг ответа по состоянию пользователя."""
-
-    _step: str
-    _origin: TgAnswer
-    _redis: Redis
-    _logger: LogSink
+class CachedUserState(UserState):
+    """Кэширующий декоратор."""
 
     @override
-    async def build(self, update: Update) -> list[httpx.Request]:
-        """Сборка ответа.
+    def __init__(self, origin: UserState) -> None:
+        """Ctor.
 
-        :param update: Update
-        :return: list[httpx.Request]
+        :param origin: UserState
         """
-        step = await RedisUserState(self._redis, TgChatId(update), self._logger).step()
-        if step.value != self._step:
-            return []
-        return await self._origin.build(update)
+        self._origin = origin
+        self._cache: UserStep | None = None
+
+    @override
+    async def step(self) -> UserStep:
+        """Состояние пользователя.
+
+        :return: UserStep
+        """
+        if self._cache:
+            return self._cache
+        self._cache = await self._origin.step()
+        return self._cache
+
+    @override
+    async def change_step(self, step: UserStep) -> None:
+        """Изменение, состояние пользователя.
+
+        :param step: UserStep
+        """
+        await self._origin.change_step(step)
+        self._cache = step
