+++ src/tests/unit/test_user_state.py
@@ -25,7 +25,8 @@ from fakeredis import aioredis
 
 from app_types.intable import FkIntable
 from app_types.logger import FkLogSink
-from services.user_state import RedisUserState, UserStep
+from srv.users.redis_user_state import RedisUserState
+from srv.users.user_step import UserStep
 
 
 @pytest.fixture()
